Rename storage for preferences
