+++ extension/content/.common/services/preferences-storage.ts
